+++ src/components/activity/WeeklyResetManager.tsx
@@ -1,10 +1,10 @@
 "use client";
 
-import { useWeeklyReset } from "@/hooks/useWeeklyReset";
+import { useWeeklyArchiveAndReset } from "@/hooks/useWeeklyArchiveAndReset";
 
 // Component to manage the weekly reset of activities
 export default function WeeklyResetManager() {
-  useWeeklyReset();
+  useWeeklyArchiveAndReset();
 
   // This component doesn't render anything
   return null;