+++ src/components/containers/ActivityDetailsContainer.tsx
@@ -1,19 +1,39 @@
 "use client";
 
-import { useParams } from "next/navigation";
+import {
+  useParams,
+  usePathname,
+  useRouter,
+  useSearchParams,
+} from "next/navigation";
 import { useEffect, useState } from "react";
 
 import { ValidatedActivity } from "@/lib/validation/activity/activity.types";
 import NoActivity from "../activity/NoActivity";
 import ActivityDetailsCard from "../cards/ActivityDetailsCard";
 import { getCurrentWeekKey } from "@/lib/utils/date";
-import { useWeeklySummary } from "@/hooks/useWeeklyHistory";
+import { useWeeklySummary } from "@/hooks/useWeeklySummary";
+import { validateWeekKey } from "@/lib/validation/history/history.validators";
+import { useActivityStore } from "@/lib/useActivityStore";
+import { ActivityDetailsCardSkeleton } from "../skeletons/components/ActivityDetailsCardSkeleton";
 
 export default function ActivityDetailsContainer() {
   const params = useParams();
+  const router = useRouter();
+  const pathname = usePathname();
+  const searchParams = useSearchParams();
+
+  // Get the activity id from the URL
+  const activityId = params.activityId as string;
+  // Get the week key from the URL and validate it
+  const weekKeyFromParams = searchParams.get("week");
+  const validatedWeekKey = validateWeekKey(weekKeyFromParams);
 
   // State to manage the selected week
-  const [selectedWeek, setSelectedWeek] = useState(getCurrentWeekKey());
+  const [selectedWeek, setSelectedWeek] = useState(
+    validatedWeekKey ?? getCurrentWeekKey()
+  );
+
   // State for the activity
   const [activity, setActivity] = useState<ValidatedActivity | null>(null);
   // State for the percentage of the progress bar
@@ -21,12 +41,12 @@ export default function ActivityDetailsContainer() {
 
   // Fetch the summary of the selected week
   const { activities } = useWeeklySummary(selectedWeek);
+  //
+  const { isRehydrated } = useActivityStore();
 
-  // Get the activity id from the URL
-  const activityId = params.activityId as string;
-
+  // Fetch the activity details from the store
   useEffect(() => {
-    // Find the activity by ID
+    // Find the activity by its ID
     const foundActivity = activities.find((act) => act.id === activityId);
 
     if (foundActivity) {
@@ -44,7 +64,18 @@ export default function ActivityDetailsContainer() {
     }
   }, [activityId, activities]);
 
-  return (
+  // Update the URL when the selected week changes
+  const handleWeekChange = (newWeekKey: string) => {
+    setSelectedWeek(newWeekKey);
+    const params = new URLSearchParams(searchParams.toString());
+    params.set("week", newWeekKey);
+
+    router.replace(`${pathname}?${params.toString()}`);
+  };
+
+  return !isRehydrated ? (
+    <ActivityDetailsCardSkeleton />
+  ) : (
     <>
       {!activity ? (
         <NoActivity />
@@ -53,7 +84,7 @@ export default function ActivityDetailsContainer() {
           activity={activity}
           percentage={percentage}
           selectedWeek={selectedWeek}
-          setSelectedWeek={setSelectedWeek}
+          setSelectedWeek={handleWeekChange}
         />
       )}
     </>