+++ src/lib/useActivityStore.ts
@@ -1,4 +1,3 @@
-// store/index.ts
 import { create } from "zustand";
 import { persist, createJSONStorage } from "zustand/middleware";
 import { ActivitySliceState } from "./validation/activity/activity.types";
@@ -7,13 +6,15 @@ import { HistorySliceState } from "./validation/history/history.types";
 import { createActivitySlice } from "./slices/activitySlice";
 import { createTimerSlice } from "./slices/timerSlice";
 import { createHistorySlice } from "./slices/historySlice";
+import { createGlobalSlice, GlobalSliceState } from "./slices/globalSlice";
 import { validateActivitiesFromStorage } from "./validation/activity/activity.validators";
 import { validateHistoryFromStorage } from "./validation/history/history.validators";
 
 // Combine all slices into one store
 export type StoreState = ActivitySliceState &
   TimerSliceState &
-  HistorySliceState;
+  HistorySliceState &
+  GlobalSliceState;
 
 /**
  * Creates the Zustand store for managing activities and timers
@@ -26,15 +27,8 @@ export const useActivityStore = create<StoreState>()(
         ...createActivitySlice(...a),
         ...createTimerSlice(...a),
         ...createHistorySlice(...a),
+        ...createGlobalSlice(...a),
       };
-
-      // Override the resetWeeklyProgress method to save the week to history before resetting
-      const originalResetWeeklyProgress = store.resetWeeklyProgress;
-      store.resetWeeklyProgress = () => {
-        store.saveWeekToHistory();
-        originalResetWeeklyProgress();
-      };
-
       return store;
     },
     {
@@ -56,6 +50,7 @@ export const useActivityStore = create<StoreState>()(
               state.weeklyHistory
             );
           }
+          state.setIsRehydrated(true);
         }
       },
     }